more fixes

+++ integration-tests/src/test/java/io/apicurio/tests/ConfluentBaseIT.java
@@ -68,16 +68,20 @@ public abstract class ConfluentBaseIT extends ApicurioRegistryBaseIT {
     }
 
     protected void clearAllConfluentSubjects() throws IOException, RestClientException {
-        for (String confluentSubject : confluentService.getAllSubjects()) {
-            logger.info("Deleting confluent schema {}", confluentSubject);
-            try {
-                confluentService.deleteSubject(confluentSubject);
-            } catch (RestClientException e) {
-                if (e.getStatus() == 404) {
-                    //subjects may be already deleted
-                    continue;
+        if (confluentService != null && confluentService.getAllSubjects() != null) {
+            for (String confluentSubject : confluentService.getAllSubjects()) {
+                if (confluentSubject != null) {
+                    logger.info("Deleting confluent schema {}", confluentSubject);
+                    try {
+                        confluentService.deleteSubject(confluentSubject);
+                    } catch (RestClientException e) {
+                        if (e.getStatus() == 404) {
+                            //subjects may be already deleted
+                            continue;
+                        }
+                        throw e;
+                    }
                 }
-                throw e;
             }
         }
     }
